requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
Pillow
psycopg2-binary  # If PostgreSQL
requests
python-decouple
PyJWT
